pyproject.toml
[project]
name = "mutemepy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["hidapi>=0.14.0.post2"]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest>=8.3.3",
]
